Remove the development config map
